cleanup framework stuff and typing
